+++ src/Features/StatefulMap/actionTypes.ts
@@ -1,4 +1,0 @@
-/**
- * Map state action types
- */
-export const SET_VIEW = "map_state/view/set"