+++ src/Features/Units/actionTypes.ts
@@ -1,5 +1,0 @@
-/**
- * Unit system management action types
- */
-
-export const UNIT_SWITCH = "unit/switch"