+++ src/Features/Units/index.ts
@@ -1,4 +1,4 @@
-export { unitReducer } from "./reducer"
+export { unitReducer } from "./slice"
 
 export { UnitSelector } from "./UnitSelector"
 export { UnitSystemConnector } from "./Connector"